refactor(front-api): relocate config properties and scan
