+++ front-api/src/main/java/org/open4goods/nudgerfrontapi/config/properties/ApiProperties.java
@@ -1,24 +1,22 @@
-package org.open4goods.nudgerfrontapi.config;
+package org.open4goods.nudgerfrontapi.config.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
 
 /**
  * Configuration properties for front cache settings.
  */
-@Component
 @ConfigurationProperties(prefix = "front")
 public class ApiProperties {
 
     /** The resource provider root path (where static resources will be hosted) */
     private String resourceProviderRootPath;
 
-	public String getResourceProviderRootPath() {
-		return resourceProviderRootPath;
-	}
+    public String getResourceProviderRootPath() {
+        return resourceProviderRootPath;
+    }
 
-	public void setResourceProviderRootPath(String resourceProviderRootPath) {
-		this.resourceProviderRootPath = resourceProviderRootPath;
-	}
+    public void setResourceProviderRootPath(String resourceProviderRootPath) {
+        this.resourceProviderRootPath = resourceProviderRootPath;
+    }
 
 }
