+++ front-api/src/main/java/org/open4goods/nudgerfrontapi/interceptor/RateLimitingInterceptor.java
@@ -4,7 +4,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.apache.commons.lang3.StringUtils;
-import org.open4goods.nudgerfrontapi.config.RateLimitProperties;
+import org.open4goods.nudgerfrontapi.config.properties.RateLimitProperties;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
