+++ front-api/src/test/java/org/open4goods/nudgerfrontapi/controller/api/TeamControllerTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.open4goods.nudgerfrontapi.config.TeamProperties;
+import org.open4goods.nudgerfrontapi.config.properties.TeamProperties;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
