+++ front-api/src/test/java/org/open4goods/nudgerfrontapi/interceptor/RateLimitingInterceptorTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.open4goods.nudgerfrontapi.config.RateLimitProperties;
+import org.open4goods.nudgerfrontapi.config.properties.RateLimitProperties;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.security.authentication.TestingAuthenticationToken;
